requirements.txt
Click==7.0
dash==1.17.0
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
Flask==1.1.1
Flask-Compress==1.4.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
plotly==4.12.0
PyYAML==5.1.2
retrying==1.3.3
six==1.12.0
Werkzeug==0.15.5
pandas==01.1.2
nltk==3.4.4
textblob==0.15.2
tweepy==3.9.0
psycopg2-binary==2.8.3
Flask==1.1.1
numpy==1.19.2